+++ src/DtHub.App/Windows/ConfiguratorWindow.xaml.cs
@@ -132,14 +132,6 @@ public partial class ConfiguratorWindow : Window
 
     private void OnHide(object sender, RoutedEventArgs e) => Hide();
 
-    /// <summary>Explique comment obtenir une seconde installation du jeu.</summary>
-    private void OnCloneHelp(object sender, RoutedEventArgs e)
-    {
-        var help = AppHost.Services.GetRequiredService<CloneHelpWindow>();
-        help.Owner = this;
-        help.ShowDialog();
-    }
-
     private void OnSleepHelp(object sender, RoutedEventArgs e)
     {
         var help = AppHost.Services.GetRequiredService<SleepHelpWindow>();